Servo writes working, project configured for OpenOCD debugging using picoprobe and a second Pi Pico acting as the SWD debugger

+++ inc/regwrite.h
@@ -2,7 +2,7 @@
 #define REG_WRITE_H
 
 int reg_write(i2c_inst_t *i2c,
-              const uint addr,
+              const uint8_t addr,
               const uint8_t reg,
               uint8_t *buf,
               const uint8_t nbytes);
